requirements.txt
faiss-cpu
groq
langchain-groq
langchain
streamlit
langchain_community
python-dotenv
langchain_core